+++ app/layout.tsx
@@ -90,8 +90,54 @@ export default function RootLayout({
 }: Readonly<{
   children: React.ReactNode;
 }>) {
+  const websiteSchema = {
+    "@context": "https://schema.org",
+    "@type": "WebSite",
+    name: "Dr. Rohan Yashraj Gupta, FIA | Actuary & Data Scientist",
+    url: "https://rohanyashraj.github.io",
+    description:
+      "Explore the actuarial journey, projects, and insights of Dr. Rohan Gupta, FIA. He specializes in Data Science and General Insurance.",
+  };
+
+  const personSchema = {
+    "@context": "https://schema.org",
+    "@type": "Person",
+    name: "Dr. Rohan Yashraj Gupta",
+    alternateName: "Rohan Gupta",
+    url: "https://rohanyashraj.github.io",
+    jobTitle: "Actuary & Data Scientist", // You might want to make this more specific or add multiple
+    worksFor: {
+      "@type": "Organization",
+      name: "Bajaj Allianz General Insurance", // Replace with actual if applicable, or remove if self-employed/independent
+    },
+    alumniOf: {
+      "@type": "CollegeOrUniversity",
+      name: "Sri Sathya Sai Institute of Higher Learning",
+    },
+    sameAs: [
+      // Add links to your professional profiles
+      "https://www.linkedin.com/in/dr-rohan-yashraj-gupta/", // Example
+      "https://github.com/RohanYashraj", // Example
+      "https://twitter.com/Rohan_Yashraj", // Example
+      // Add other relevant profiles (e.g., ResearchGate, Google Scholar)
+    ],
+    image: "https://rohanyashraj.github.io/profile.jpg", // Ensure this image is optimized!
+    description:
+      "Qualified General Insurance Actuary (FIA, FIAI) and Data Scientist with a PhD, specializing in fraud detection, predictive modeling, and technology transformation in the insurance sector.",
+  };
+
   return (
     <html lang="en">
+      <head>
+        <script
+          type="application/ld+json"
+          dangerouslySetInnerHTML={{ __html: JSON.stringify(websiteSchema) }}
+        />
+        <script
+          type="application/ld+json"
+          dangerouslySetInnerHTML={{ __html: JSON.stringify(personSchema) }}
+        />
+      </head>
       <body
         className={`${raleway.variable} antialiased bg-bodyColour text-white/80`}
       >
