+++ components/Logo.tsx
@@ -11,7 +11,7 @@ const Logo = ({ className, title, subtitle }: Props) => {
   return (
     <div className="text-2xl group">
       <Link href={"/"}>
-        <h2
+        <span
           className={cn(
             "font-semibold tracking-wide hover:text-hoverColor hoverEffect",
             className
@@ -22,7 +22,7 @@ const Logo = ({ className, title, subtitle }: Props) => {
           <span className="text-lightSky group-hover:text-white hoverEffect">
             {subtitle}
           </span>
-        </h2>
+        </span>
       </Link>
     </div>
   );
